apps/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7 
sqlalchemy==2.0.21
asyncpg==0.29.0
alembic==1.12.1
pydantic>=2.0.0,<3.0.0 
fastapi-users[sqlalchemy,jwt]==12.1.3
passlib[bcrypt]==1.7.4 
bcrypt==3.2.2
fastapi-users-db-sqlalchemy==7.0.0 
boto3==1.34.48 
httpx>=0.28.1,<1.0.0
google-genai==1.19.0
gTTS==2.5.1
requests>=2.31.0
uvloop==0.20.0; sys_platform != "win32"
azure-cognitiveservices-speech==1.35.0
redis==5.0.7
pdfminer.six==20231228
python-docx==0.8.11
cryptography>=41.0.0
psutil>=5.9.0
openai>=1.102.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-instrumentation-starlette>=0.47b0
opentelemetry-instrumentation-httpx>=0.47b0
opentelemetry-instrumentation-sqlalchemy>=0.47b0
opentelemetry-instrumentation-redis>=0.47b0
sentry-sdk>=2.10.0
prometheus-client>=0.20.0
opentelemetry-exporter-otlp-proto-http>=1.26.0
